refactor: Migrate profile template in a Gradle module.
