+++ app/build.gradle.kts
@@ -80,6 +80,7 @@ dependencies {
     implementation(project(":demos:moviesapp:app"))
     implementation(project(":templates:onboarding"))
     implementation(project(":templates:paymentcard"))
+    implementation(project(":templates:profile"))
     implementation(project(":components:fab"))
     implementation(project(":components:charts"))
     implementation(project(":components:tags"))
