+++ settings.gradle.kts
@@ -25,3 +25,4 @@ include(":animations:canvas")
 include(":components:comingsoon")
 include(":templates:onboarding")
 include(":templates:paymentcard")
+include(":templates:profile")
