+++ templates/profile/src/main/java/com/guru/composecookbook/profile/MyPhotosSection.kt
@@ -1,8 +1,6 @@
-package com.guru.composecookbook.ui.templates.profile
+package com.guru.composecookbook.profile
 
 import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.Divider
@@ -12,11 +10,8 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
-import com.guru.composecookbook.R
-import com.guru.composecookbook.theme.green200
-import com.guru.composecookbook.theme.green700
+import com.guru.composecookbook.data.R
 import com.guru.composecookbook.theme.typography
 
 @Composable
