+++ templates/profile/src/main/res/values/strings.xml
@@ -1,0 +1,4 @@
+<resources>
+    <string name="about_me">I am an enthusiastic Developer. I love to create things that make the world a better place. I like to work where technology is used to improve core components of life like Health, Education and Environment.</string>
+    <string name="about_project">This project was started because of my curiosity about declarative UIs.I have been tracking Jetpack Compose long before alpha release and trying to understand how it is gonna fit in our current existing system. So this is an attempt to showcase how it can be used to make android development more fun, easy and interesting.</string>
+</resources>
